use long as id in repository

+++ src/main/java/my/photomanager/photo/album/PhotoAlbumRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PhotoAlbumRepository extends JpaRepository<PhotoAlbum, Integer> {
+public interface PhotoAlbumRepository extends JpaRepository<PhotoAlbum, Long> {
 
 	Optional<PhotoAlbum> findByName(@NonNull String name);
 }
